# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PC Builder API",
    "contact": {
      "name": "Juozas Petryla"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/builds/{build_id}/reviews": {
      "get": {
        "tags": [
          "Atsiliepimai"
        ],
        "summary": "Gauti komplekto atsiliepimų sąrašą",
        "description": "Hierarchinis sąrašo metodas: grąžina visus konkrečiam komplektui priklausančius atsiliepimus.",
        "operationId": "listBuildReviews",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewRead"
                  },
                  "title": "Response Listbuildreviews"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Komplektas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Atsiliepimai"
        ],
        "summary": "Įvertinti ir pakomentuoti komplektą",
        "description": "Prideda 1–5 balų įvertinimą ir tekstinį komentarą prie komplekto.",
        "operationId": "createBuildReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Komplektas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Neteisingas įvertinimas arba komentaras."
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}": {
      "get": {
        "tags": [
          "Atsiliepimai"
        ],
        "summary": "Gauti atsiliepimą",
        "description": "Grąžina vieną atsiliepimą pagal jo identifikatorių.",
        "operationId": "getReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Atsiliepimas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Atsiliepimai"
        ],
        "summary": "Atnaujinti atsiliepimą",
        "description": "Pakeičia savo atsiliepimo įvertinimą ir komentarą; autorius nesikeičia.",
        "operationId": "replaceReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Atsiliepimas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Neteisingas įvertinimas arba komentaras."
          }
        }
      },
      "delete": {
        "tags": [
          "Atsiliepimai"
        ],
        "summary": "Pašalinti atsiliepimą",
        "description": "Autorius šalina savo atsiliepimą matomame komplekte; moderatorius ir admin gali šalinti svetimą atsiliepimą viešame komplekte.",
        "operationId": "deleteReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Atsiliepimas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCreate": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "examples": [
              5
            ]
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rating",
          "comment"
        ],
        "title": "ReviewCreate"
      },
      "ReviewLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "build": {
            "type": "string",
            "title": "Build"
          }
        },
        "type": "object",
        "required": [
          "self",
          "build"
        ],
        "title": "ReviewLinks"
      },
      "ReviewRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "build_id": {
            "type": "integer",
            "title": "Build Id"
          },
          "author_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "links": {
            "$ref": "#/components/schemas/ReviewLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "build_id",
          "author_id",
          "author_name",
          "rating",
          "comment",
          "created_at",
          "links"
        ],
        "title": "ReviewRead"
      },
      "ReviewReplace": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "examples": [
              5
            ]
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rating",
          "comment"
        ],
        "title": "ReviewReplace",
        "description": "Visas atsiliepimo turinys, naudojamas PUT operacijoje."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}